self replication of the script to be ran multiple times without user having to provide inputs again
Pythoncopy is a template that gets put in the users folder. With the data saved in the users file from the inputs, the script will actually replicate the users inputs. This is intended to allow further automatization through task scheduler or other tools

delft_to_schiphol.py is a working example of this.

replicated_script.py currently stores the users inputs and gets called when the other python script in the folder runs.